Remove awstesting dependancy on API generator
Removes the dependancies on the API generator from awstesting. This
ensures that the API generator does not leak into the SDK or the SDK
testing.
